# gateway/build/src/server/swagger.json
{
  "openapi": "3.0.0",
  "components": {
...
    "schemas": {
...
      "Invoice": {
        "properties": {
          "id": {
            "type": "string"
          },
          "number": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["SALES", "PURCHASE"]
          },
          "date": {
            "type": "string"
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "type": "string"
          }
        },
        "required": ["id", "number", "type", "date", "total", "status"],
        "type": "object",
        "additionalProperties": false
      },
      "InvoiceCreateRequest": {
        "properties": {
          "type": {
            "type": "string",
            "enum": ["SALES", "PURCHASE"]
          },
          "customerId": {
            "type": "string"
          },
          "supplierId": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "dueDate": {
            "type": "string"
          },
          "items": {
            "items": {
              "properties": {
                "unitPrice": {
                  "type": "number",
                  "format": "double"
                },
                "quantity": {
                  "type": "number",
                  "format": "double"
                },
                "description": {
                  "type": "string"
                }
              },
              "required": ["unitPrice", "quantity", "description"],
              "type": "object"
            },
            "type": "array"
          }
        },
        "required": ["type", "date", "items"],
        "type": "object",
        "additionalProperties": false
      },
...
      "Employee": {
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "position": {
            "type": "string"
          },
          "salary": {
            "type": "number",
            "format": "double"
          },
          "startDate": {
            "type": "string"
          }
        },
        "required": ["id", "firstName", "lastName", "email", "position", "salary", "startDate"],
        "type": "object",
        "additionalProperties": false
      },
      "EmployeeCreateRequest": {
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "position": {
            "type": "string"
          },
          "salary": {
            "type": "number",
            "format": "double"
          },
          "startDate": {
            "type": "string"
          }
        },
        "required": ["firstName", "lastName", "email", "position", "salary", "startDate"],
        "type": "object",
        "additionalProperties": false
      },
      "Deal": {
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "value": {
            "type": "number",
            "format": "double"
          },
          "stage": {
            "type": "string",
            "enum": ["PROSPECT", "QUALIFIED", "PROPOSAL", "NEGOTIATION", "CLOSED_WON", "CLOSED_LOST"]
          },
          "contactId": {
            "type": "string"
          },
          "expectedCloseDate": {
            "type": "string"
          },
          "probability": {
            "type": "number",
            "format": "double"
          }
        },
        "required": ["id", "title", "value", "stage", "contactId", "expectedCloseDate", "probability"],
        "type": "object",
        "additionalProperties": false
      },
      "DealCreateRequest": {
        "properties": {
          "title": {
            "type": "string"
          },
          "value": {
            "type": "number",
            "format": "double"
          },
          "contactId": {
            "type": "string"
          },
          "expectedCloseDate": {
            "type": "string"
          },
          "probability": {
            "type": "number",
            "format": "double"
          }
        },
        "required": ["title", "value", "contactId", "expectedCloseDate", "probability"],
        "type": "object",
        "additionalProperties": false
      },
      "Contact": {
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "company": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["LEAD", "CUSTOMER", "VENDOR"]
          },
          "status": {
            "type": "string",
            "enum": ["ACTIVE", "INACTIVE"]
          }
        },
        "required": ["id", "name", "email", "phone", "type", "status"],
        "type": "object",
        "additionalProperties": false
      },
      "ContactCreateRequest": {
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "company": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["LEAD", "CUSTOMER", "VENDOR"]
          }
        },
        "required": ["name", "email", "phone", "type"],
        "type": "object",
        "additionalProperties": false
      },
...
      "Account": {
        "properties": {
          "id": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["ASSET", "LIABILITY", "EQUITY", "REVENUE", "EXPENSE"]
          },
          "balance": {
            "type": "number",
            "format": "double"
          },
          "parentId": {
            "type": "string"
          }
        },
        "required": ["id", "code", "name", "type", "balance"],
        "type": "object",
        "additionalProperties": false
      },
      "AccountCreateRequest": {
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": ["ASSET", "LIABILITY", "EQUITY", "REVENUE", "EXPENSE"]
          },
          "parentId": {
            "type": "string"
          }
        },
        "required": ["code", "name", "type"],
        "type": "object",
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "jwt": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "description": "JWT authorization header using the Bearer scheme. Format: Authorization: Bearer {token}"
      }
    }
  },
  "info": {
    "title": "Safee Analytics API",
    "version": "1.0.0",
    "description": "API documentation for Safee Analytics SaaS platform",
    "license": {
      "name": "MIT"
    },
    "contact": {
      "name": "Safee Analytics",
      "email": "support@safee-analytics.com"
    }
  },
  "paths": {
...
    "/invoices": {
      "get": {
        "operationId": "GetInvoices",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "limit": {
                      "type": "number",
                      "format": "double"
                    },
                    "page": {
                      "type": "number",
                      "format": "double"
                    },
                    "total": {
                      "type": "number",
                      "format": "double"
                    },
                    "invoices": {
                      "items": {
                        "$ref": "#/components/schemas/Invoice"
                      },
                      "type": "array"
                    }
                  },
                  "required": ["limit", "page", "total", "invoices"],
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": ["Invoices"],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "format": "double",
              "type": "number"
            }
          }
        ]
      },
      "post": {
        "operationId": "CreateInvoice",
        "responses": {
          "201": {
            "description": "Invoice created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          }
        },
        "tags": ["Invoices"],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceCreateRequest"
              }
            }
          }
        }
      }
    },
    "/hisabiq/invoices": {
      "get": {
        "operationId": "GetInvoices",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "limit": {
                      "type": "number",
                      "format": "double"
                    },
                    "page": {
                      "type": "number",
                      "format": "double"
                    },
                    "total": {
                      "type": "number",
                      "format": "double"
                    },
                    "invoices": {
                      "items": {
                        "$ref": "#/components/schemas/Invoice"
                      },
                      "type": "array"
                    }
                  },
                  "required": ["limit", "page", "total", "invoices"],
                  "type": "object"
                }
              }
            }
          }
        },
        "description": "Get all invoices for the authenticated user's organization",
        "tags": ["Hisabiq (Accounting)"],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "format": "double",
              "type": "number"
            }
          }
        ]
      },
      "post": {
        "operationId": "CreateInvoice",
        "responses": {
          "201": {
            "description": "Invoice created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          }
        },
        "description": "Create a new invoice",
        "tags": ["Hisabiq (Accounting)"],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceCreateRequest"
              }
            }
          }
        }
      }
    },
    "/hisabiq/accounts": {
      "get": {
        "operationId": "GetAccounts",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "properties": {
                      "parentId": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "code": {
                        "type": "string"
                      },
                      "id": {
                        "type": "string"
                      }
                    },
                    "required": ["type", "name", "code", "id"],
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            }
          }
        },
        "description": "Get chart of accounts",
        "tags": ["Hisabiq (Accounting)"],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": []
      }
    },
...
    "/employees": {
      "get": {
        "operationId": "GetEmployees",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "limit": {
                      "type": "number",
                      "format": "double"
                    },
                    "page": {
                      "type": "number",
                      "format": "double"
                    },
                    "total": {
                      "type": "number",
                      "format": "double"
                    },
                    "employees": {
                      "items": {
                        "$ref": "#/components/schemas/Employee"
                      },
                      "type": "array"
                    }
                  },
                  "required": ["limit", "page", "total", "employees"],
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": ["HR & Payroll"],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "format": "double",
              "type": "number"
            }
          }
        ]
      },
      "post": {
        "operationId": "CreateEmployee",
        "responses": {
          "201": {
            "description": "Employee created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              }
            }
          }
        },
        "tags": ["HR & Payroll"],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeCreateRequest"
              }
            }
          }
        }
      }
    },
    "/deals": {
      "get": {
        "operationId": "GetDeals",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "limit": {
                      "type": "number",
                      "format": "double"
                    },
                    "page": {
                      "type": "number",
                      "format": "double"
                    },
                    "total": {
                      "type": "number",
                      "format": "double"
                    },
                    "deals": {
                      "items": {
                        "$ref": "#/components/schemas/Deal"
                      },
                      "type": "array"
                    }
                  },
                  "required": ["limit", "page", "total", "deals"],
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": ["CRM"],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "stage",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "operationId": "CreateDeal",
        "responses": {
          "201": {
            "description": "Deal created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deal"
                }
              }
            }
          }
        },
        "tags": ["CRM"],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DealCreateRequest"
              }
            }
          }
        }
      }
    },
    "/contacts": {
      "get": {
        "operationId": "GetContacts",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "limit": {
                      "type": "number",
                      "format": "double"
                    },
                    "page": {
                      "type": "number",
                      "format": "double"
                    },
                    "total": {
                      "type": "number",
                      "format": "double"
                    },
                    "contacts": {
                      "items": {
                        "$ref": "#/components/schemas/Contact"
                      },
                      "type": "array"
                    }
                  },
                  "required": ["limit", "page", "total", "contacts"],
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": ["CRM"],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "operationId": "CreateContact",
        "responses": {
          "201": {
            "description": "Contact created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          }
        },
        "tags": ["CRM"],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactCreateRequest"
              }
            }
          }
        }
      }
    },
...
    "/accounts": {
      "get": {
        "operationId": "GetAccounts",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Account"
                  },
                  "type": "array"
                }
              }
            }
          }
        },
        "tags": ["Accounting"],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "operationId": "CreateAccount",
        "responses": {
          "201": {
            "description": "Account created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          }
        },
        "tags": ["Accounting"],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountCreateRequest"
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}